[MAT-290] Rename terminals to exprtrees in preparation for putting all the tree nodes in there.
