+++ src/librdag/exprtree.py
